+++ docs/CHANGELOG.md
@@ -113,6 +113,10 @@ title: Changelog
 
     *Joel Hawksley*
 
+* Add support for multiple templates.
+
+    *Rob Sterner*, *Joel Hawksley*
+
 ## 2.36.0
 
 * Add `slot_type` helper method.